relocate Bodu.Security.Cryptography-only contracts and KATs out of Bodu.Test
Five contract bases and five KAT records (incl. one enum) were consumed
only by Bodu.Security.Cryptography.Test, so they move alongside their
consumers per the single-consumer rule.

Contracts moved to Bodu.Security.Cryptography/test/Security.Cryptography/
Contracts/ (namespace Bodu.Security.Cryptography.Contracts to match
existing concrete subclasses):

- AeadContractTests<TAead>
- BlockCipherContractTests<TCipher>
- BlockCipherModeContractTests<TCipher>
- CryptoHashContractTests<THash>
- CryptoTransformContractTests<TTransform>

KATs moved to Bodu.Security.Cryptography/test/Infrastructure/ (namespace
Bodu.Security.Cryptography.Infrastructure to match existing
AeadKnownAnswerVector):

- AeadKat
- AeadTamperKat (+ AeadTamperKind enum)
- BlockCipherKat
- BlockCipherModeKat
- CryptoHashKat

Subclass `using` updates (18 files in
Security.Cryptography/Contracts/):
- Removed unused `using Bodu.Test.Contracts;` (now same namespace as
  the bases).
- Replaced `using Bodu.Test.Kat;` with
  `using Bodu.Security.Cryptography.Infrastructure;`.

Duplicate KATs investigated but NOT consolidated — on inspection they
are not duplicates of the relocated records, they are a different
abstraction tied to the BlockCipherSpecification /
HashAlgorithmKnownAnswers framework with richer fields (nullable Key
defaulting to TestKey, Profile tags, hex-encoded ExpectedHex, NIST
Count/Source). Consolidating them would require rewriting the
SymmetricAlgorithmTests / BlockCipherTransformTests harness which is
out of scope:
- BlockCipherKnownAnswer
- HashAlgorithmKnownAnswer
- KeyedHashAlgorithmKnownAnswer
- AeadKnownAnswerVector

Verified: dotnet build passes 0 errors;
dotnet test --settings bvt.runsettings passes 15299/15299 (369
pre-existing skips).
