+++ Bodu.Security.Cryptography/test/Infrastructure/CryptoHashKat.cs
@@ -1,10 +1,12 @@
-﻿// ---------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------
 // <copyright file="CryptoHashKat.cs" company="Bodu Pty. Ltd.">
 //     Copyright (c) Bodu Pty. Ltd.. All rights reserved.
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-namespace Bodu.Test.Kat;
+using Bodu.Test.Kat;
+
+namespace Bodu.Security.Cryptography.Infrastructure;
 
 /// <summary>
 /// Represents a known-answer test row for a cryptographic hash, keyed hash, or XOF algorithm. Carries the raw input,
