+++ Bodu.Security.Cryptography/test/Security.Cryptography/Contracts/SkipjackBlockCipherModeContractTests.cs
@@ -5,8 +5,7 @@
 // ---------------------------------------------------------------------------------------------------------------
 
 using System.Security.Cryptography;
-using Bodu.Test.Contracts;
-using Bodu.Test.Kat;
+using Bodu.Security.Cryptography.Infrastructure;
 
 namespace Bodu.Security.Cryptography.Contracts;
